+++ packages/react-core/src/components/atoms/Icon/Icon.tsx
@@ -1,15 +1,14 @@
 import React, { FC } from 'react';
 import { StyleProp, TextStyle, Text } from 'react-native';
-import { useTheme } from '@emotion/react';
 import {
   ColorGradationType,
   ColorType,
   FontColorType,
   IconSizeType,
   IconType,
-  ThemeProp,
+  useTheme,
 } from '@tecsinapse/react-core';
-import { getIconColor, getIconComponent } from './helpers';
+import { getIconColor, getIconComponent } from './functions';
 
 export interface IconProps {
   /**  Name of the icon. You must use the same icons from react-native-vector-icons */
@@ -38,13 +37,13 @@ const Icon: FC<IconProps> = ({
   colorTone = 'medium',
   ...rest
 }) => {
-  const theme = useTheme() as ThemeProp;
+  const theme = useTheme();
   const color = getIconColor(colorVariant, colorTone, fontColor, theme);
-  const RNVIcon = getIconComponent(type, size);
+  const RNVIcon = getIconComponent(type);
 
   return (
     <React.Suspense fallback={<Text> </Text>}>
-      <RNVIcon {...rest} style={style} name={name} color={color} />
+      <RNVIcon {...rest} style={style} name={name} color={color} size={size} />
     </React.Suspense>
   );
 };